% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celegansConverter.R
\name{celegansConverter}
\alias{celegansConverter}
\title{Convert C. elegans gene names to human gene names}
\usage{
celegansConverter(genelist, colname)
}
\arguments{
\item{genelist}{A dataframe object}

\item{colname}{Name of the column containing gene synonyms to be converted to gene names}
}
\value{
dataframe object
}
\description{
This function converts a column containing C. elegans gene names within a data frame
to human gene names. If the column contains a mix of C. elegans and Human gene names,
only gene synonyms will be converted to gene names and gene names will remain the same.
}
