% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mousegnameConverter.R
\name{mousegnameConverter}
\alias{mousegnameConverter}
\title{Convert mouse gene names to human NCBI gene names}
\usage{
mousegnameConverter(genelist, colname)
}
\arguments{
\item{genelist}{A dataframe object}

\item{colname}{Name of the column containing mouse gene names to be converted to human gene names}
}
\value{
dataframe object
}
\description{
This function converts a column containing mouse gene names within a data frame
to human NCBI gene names. If the column contains a mix of mouse gene names and human gene names,
only mouse gene names will be converted to human gene names and mouse gene names will remain the same.
}
