% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mousesynonymConverter.R
\name{mousesynonymConverter}
\alias{mousesynonymConverter}
\title{Convert mouse gene synonyms to NCBI mouse gene names}
\usage{
mousesynonymConverter(genelist, colname)
}
\arguments{
\item{genelist}{A dataframe object}

\item{colname}{Name of the column containing mouse gene synonyms to be converted to mouse gene names}
}
\value{
dataframe object
}
\description{
This function converts a column containing mouse gene synonyms within a data frame
to mouse NCBI gene names. If the column contains a mix of mouse gene names and mouse gene synonyms,
only mouse gene synonyms will be converted to mouse gene names and mouse gene names will remain the same.
}
